Trekker trygdetid-dtoer ut i saksbehandling-common
